Rewritten section 5.1
It revises @fergunet text, eliminates the section and merges it into
the introduction, and revises and adds bibliography. In general,
rewriting should not be limited to say in different words but to
actually adapt to the new focus of the paper (metaheuristics) as well
as update references (and how you mention them). Refs #58 but also
updates bibliography, checking #35.